Move color components to separate package
