+++ packages/components-color/.storybook/preview-head.html
@@ -1,0 +1,5 @@
+<style>
+  .sb-show-main {
+    padding: 0 !important; /* stylelint-disable-line declaration-no-important */
+  }
+</style>
