+++ packages/components-color/.storybook/preview.js
@@ -24,18 +24,7 @@
 
  */
 
-export * from './FieldCheckbox'
-export * from './FieldCheckboxGroup'
-export * from './FieldChips'
-export * from './FieldRadio'
-export * from './FieldRadioGroup'
-export * from './FieldRangeSlider'
-export * from './FieldSelect'
-export * from './FieldSelectMulti'
-export * from './FieldSlider'
-export * from './FieldText'
-export * from './FieldTime'
-export * from './FieldTimeSelect'
-export * from './FieldTextArea'
-export * from './FieldToggleSwitch'
-export * from './Field'
+import { addDecorator } from '@storybook/react'
+import { componentsDecorator } from '../../storybook-config/src/componentsDecorator'
+
+addDecorator(componentsDecorator)
