+++ packages/components-color/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@looker/components",
+  "name": "@looker/components-color",
   "license": "MIT",
   "version": "0.9.29",
   "main": "lib/index.cjs.js",
@@ -22,20 +22,12 @@
     "access": "public"
   },
   "dependencies": {
-    "@looker/components-providers": "^0.9.29",
+    "@looker/components": "0.9.29",
     "@looker/design-tokens": "^0.9.27",
-    "@looker/icons": "^0.9.26",
-    "@popperjs/core": "^2.5.4",
-    "date-fns": "^2.16.1",
-    "date-fns-tz": "^1.0.12",
-    "hotkeys-js": "^3.8.1",
+    "d3-color": "^2.0.0",
+    "d3-hsv": "^0.1.0",
     "lodash": "^4.17.20",
-    "react-day-picker": "^7.4.8",
-    "react-hotkeys-hook": "^2.4.0",
-    "react-resize-detector": "^5.2.0",
-    "resize-observer-polyfill": "^1.5.1",
-    "styled-system": "^5.1.5",
-    "uuid": "^8.3.2"
+    "styled-system": "^5.1.5"
   },
   "devDependencies": {
     "@looker/components-test-utils": "^0.9.29",
@@ -44,12 +36,8 @@
     "@testing-library/jest-dom": "^5.11.6",
     "@testing-library/react": "^11.2.2",
     "@testing-library/user-event": "^12.5.0",
-    "@types/lodash": "^4.14.167",
-    "@types/react": "^16.9.56",
-    "@types/react-dom": "^16.9.9",
-    "@types/react-resize-detector": "^5.0.0",
-    "@types/styled-components": "^5.1.5",
-    "@types/uuid": "^8.3.0",
+    "@types/d3-color": "^2.0.1",
+    "@types/d3-hsv": "^0.1.4",
     "csstype": "^3.0.5",
     "jest-styled-components": "^7.0.3",
     "react": "^16.14.0",
@@ -63,6 +51,5 @@
     "react-dom": "^16.11",
     "react-is": "^16.11",
     "styled-components": "^5"
-  },
-  "gitHead": "62febc162e3e45124f403e2c84fc741f68fe6714"
+  }
 }
