+++ packages/components-color/src/InputColor/utils/named_colors.ts
