+++ packages/components-color/src/Swatch/Swatch.test.tsx
