+++ packages/components-color/src/Swatch/index.ts
