+++ packages/components/src/Form/Inputs/index.ts
@@ -30,7 +30,6 @@ export * from './height'
 export * from './InlineInputText'
 export * from './InlineTextArea'
 export * from './InputChips'
-export * from './InputColor'
 export * from './InputFilters'
 export * from './InputHidden'
 export * from './InputSearch'
